audio recorder: show elapsed time during recording
